chart: move autofs root to hostPath
This commit makes it so that the nodeplugin shares its internal
/cvmfs autofs root via a hostPath instead of an emptyDir.

kubelet periodically scrapes metrics from all emptyDir volumes by
walking trough them recursively, stat()-ing each item. This however
resets the autofs mount expiry timeout, making it impossible to
automatically unmount unused CVMFS repos. The workaround for this is
to share /cvmfs via a hostPath which is not included by the metrics,
avoiding the issue.