automount: shut down automount daemon with SIGKILL
automount daemon unmounts the autofs root in /cvmfs upon
receiving SIGTERM. This makes it impossible to reconnect
the daemon to the mount later, so all consumer Pods will
loose their mounts CVMFS, without the possibility of restoring
them (unless these Pods are restarted too). The implication
is that the nodeplugin is just being restarted, and will be
needed again.

SIGKILL is handled differently in automount, as this forces
the daemon to skip the cleanup at exit, leaving the autofs
mount behind and making it possible to reconnect to it later.
We make a use of this, and unless the admin doesn't explicitly
ask for cleanup with AUTOFS_TRY_CLEAN_AT_EXIT env var, no cleanup
is done.