automount-reconciler: consult only mount table and cvmfs_talk
Previously, after checking mount table and cvmfs_talk, the reconciler
would also consult stat(). This has proven to be unreliable due to stale
caches, and then making wrong decisions when to reconcile a mount. This
resulted in reconcile delay that's larger than necessary.

This commit removes the stat() call and makes the decision solely on
results from cvmfs_talk as this seems sufficient.