sharedmount: create metadata dir only when running sharedmount-runner
The directory was created in all cvmfs-csi components (including
controllerplugin), because the MkdirAll was called in init() function.
This breaks deployments that don't run controllerplugin as root.
It doesn't have perms to create the directory, and init() exits
with panic().

This commit fixes that by separating the directory creation into
another function.