Add mount override for /etc/cvmfs/default.d/90-local.conf (#148)
The default CVMFS package installs /etc/cvmfs/default.d/50-cern.conf
which defines the config repo.  The only place that a config repo can defined
is default.local or in default.d directory, so to change the config repo (for
example, to use the OSG config repo) you must add a new file, to be read
after 50-cern.conf to the default.d directory.